+++ SIL.Archiving/ArchivingDlgViewModel.cs
@@ -224,7 +224,9 @@ namespace SIL.Archiving
 		/// <param name="title">Title of the submission</param>
 		/// <param name="id">Identifier (used as filename) for the package being created</param>
 		/// <param name="setFilesToArchive">Delegate to request client to call methods to set
-		/// which files should be archived (this is deferred to allow display of progress message)</param>
+		/// which files should be archived (this is deferred to allow display of progress
+		/// message). Clients will normally do this by calling AddFileGroup one or more times.
+		/// </param>
 		/// ------------------------------------------------------------------------------------
 		protected ArchivingDlgViewModel(string appName, string title, string id,
 			Action<ArchivingDlgViewModel, CancellationToken> setFilesToArchive)